% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_adjusting.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{Confidence Intervals}
\usage{
confidence_intervals(statistics, variances, conf_lvl = 0.95,
  conf_method = "identity")
}
\arguments{
\item{statistics}{\code{[numeric]} (mandatory, no default)

statistics for which to calculate confidence intervals}

\item{variances}{\code{[numeric]} (mandatory, no default)

variance estimates of \code{statistics} used to compute confidence intervals}

\item{conf_lvl}{\code{[numeric]} (mandatory, default \code{0.95})

confidence level of confidence intervals in \code{]0, 1[}}

\item{conf_method}{\code{[character]} (mandatory, default \code{"identity"})

see section \strong{Confidence interval methods}}
}
\value{
\code{data.table} with columns
\itemize{
\item \code{statistic}: the values you supplied via argument \code{statistics}
\item \code{variance}: the values you supplied via argument \code{variances}
\item \code{ci_lo}: lower bound of confidence interval
\item \code{ci_hi}: upper bound of confidence interval
}
}
\description{
Computes different kinds of confidence intervals given the statistics
and their variance estimates.
}
\section{Confidence interval methods}{


Currently supported confidence interval methods and their formulae:

\itemize{
 \item identity: STAT + Z * STD_ERR \item log: STAT * exp(Z * STD_ERR/STAT) \item log-log: STAT^exp(-Z * (STD_ERR/(abs(log(STAT)) * STAT)))
}


Where

\itemize{
 \item STAT is the statistic,

 \item Z is the quantile from the standard normal distribution for the 
   lower or upper bound, and

 \item STD_ERR is the standard error (square root of the variance)

}
}

