% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_adjusting.R
\name{direct_adjusted_estimates}
\alias{direct_adjusted_estimates}
\title{Direct Adjusted Estimates}
\usage{
direct_adjusted_estimates(stats_dt, specs_dt, stratum_col_nms = NULL,
  adjust_col_nms, weights)
}
\arguments{
\item{stats_dt}{\code{[data.table]} (mandatory, no default)

a \code{data.table} containing estimates and variance estimates of statistics}

\item{specs_dt}{\code{[data.table]} (mandatory, no default)

a \code{data.table} specifying names of columns in \code{stats_dt} for the estimates
(in column \code{est}) and their variance estimates (in column \code{var});
all columns of \code{stats_dt} specified in \code{specs_dt$est} are always adjusted, and
those which have corresponding non-\code{NA} values in \code{specs_dt$var} will also
have weighted variances computed, and from these the appropriate confidence
intervals; the confidence levels should be given in \code{specs_dt$conf_lvl} and
the method of computing the intervals in \code{specs_dt$conf_method};
see \code{\link{confidence_interval_methods}} for supported methods;
see also \strong{Examples}}

\item{stratum_col_nms}{\code{[NULL, character]} (optional, default \code{NULL})

names of columns in \code{stats_dt} by which statistics are stratified (and they
should be stratified by these columns after direct adjusting)}

\item{adjust_col_nms}{\code{[character]} (mandatory, no default)

names of columns in \code{stats_dt} by which statistics are currently stratified
and by which the statistics should be adjusted}

\item{weights}{\code{[double, data.table, character]}

the weights need not sum to one as this is ensured internally; you may
supply weights in one of the following ways:
\itemize{
\item \code{double}: a vector of weights can be given when adjusting by exactly
one variable (e.g. age group); the length must match the number of unique
values in the adjusting variable
\item \code{data.table}: with one or more columns with names matching to those
variables that are used to adjust estimates, and one column named
\code{weight}; e.g. \code{data.table(agegroup = 1:3, weight = c(100, 500, 400))}
\item \code{character}: a string specifying one of the weighting schemas integrated
into this package; see \code{\link{integrated_weight_schemas}}; again
this is only allowed when you are adjusting by exactly one variable
}}
}
\description{
Compute direct adjusted estimates from a table of statistics.
}
\examples{

# suppose we have poisson rates that we want to adjust for by age group.
# they are stratified by sex.
library("data.table")
set.seed(1337)

offsets <- rnorm(8, mean = 1000, sd = 100)
baseline <- 100
sex_hrs <- rep(1:2, each = 4)
age_group_hrs <- rep(c(0.75, 0.90, 1.10, 1.25), times = 2)
counts <- rpois(8, baseline * sex_hrs * age_group_hrs)

# raw estimates
my_stats <- data.table(
  sex = rep(1:2, each = 4),
  ag = rep(1:4, times = 2),
  e = counts / offsets
)
my_stats[["v"]] <- my_stats[["e"]] / offsets

# adjusted by age group
my_adj_stats <- direct_adjusted_esimates(
  stats_dt = my_stats,
  specs_dt = data.table(est = "e", var = "v", conf_lvl = 0.95),
  stratum_col_nms = "sex",
  adjust_col_nms = "ag",
  weights = c(200, 300, 400, 100)
)

}
