% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_integrated_weight_schemas.R
\docType{data}
\name{integrated_weight_schemas}
\alias{integrated_weight_schemas}
\title{Integrated Weight Schemas}
\format{\code{data.table} with columns
\itemize{
\item \code{schema_name}: name of weights schema as a string
\item \code{n_age_groups}: number of age groups in schema
\item \code{age_low}: lower bound of age group
}

All age groups are lower-bound inclusive upper-bound exclusive
(\code{[a, b[}).}
\usage{
integrated_weight_schemas
}
\description{
Weighting schemas included in this package.
}
\keyword{datasets}
