% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confidence_intervals}
\alias{confidence_intervals}
\alias{delta_method_confidence_intervals}
\title{Confidence Intervals}
\usage{
delta_method_confidence_intervals(
  statistics,
  variances,
  conf_lvl = 0.95,
  conf_method = "identity"
)
}
\arguments{
\item{statistics}{\verb{[numeric]} (no default)

Statistics for which to calculate confidence intervals.}

\item{variances}{\verb{[numeric]} (no default)

Variance estimates of \code{statistics} used to compute confidence intervals.}

\item{conf_lvl}{\verb{[numeric]} (default \code{0.95})

Confidence level of confidence intervals in \verb{]0, 1[}.}

\item{conf_method}{\verb{[character]} (default \code{"identity"})

See \strong{Details}.}
}
\description{
Functions to compute confidence intervals.
}
\section{Functions}{

\strong{directadjusting::delta_method_confidence_intervals}

\code{directadjusting::delta_method_confidence_intervals} can be used to
compute confidence intervals using the delta method. The following steps
are performed:
\itemize{
\item Collect mathematical expression \code{math} for computing the lower and upper
end of the confidence interval based on \code{conf_method}.
The following \code{conf_method} options have pre-determined math within
this R package: \code{c("identity", "log", "log-log")}.
Alternatively, \code{conf_method} can be quoted R expression --- see \code{?quote}.
In that case it will be used as-is.
\item Collect variables \code{est = statistics}, \code{var = variances}, and
\code{std_err = sqrt(variance)} into a single \code{data.table}.
\item Add \code{z = stats::qnorm(p = (1 - conf_lvl) / 2)} into the \code{data.table}.
E.g. \code{z = -1.959964} with \code{conf_lvl = 0.95}.
\item Evaluate the mathematical expression \code{math} such that the columns in the
\code{data.table} are available in evaluation. This creates column \code{ci_lo}.
\item Re-set \code{z = stats::qnorm(p = conf_lvl + (1 - conf_lvl) / 2)}.
E.g. \code{z = 1.959964} with \code{conf_lvl = 0.95}.
\item Evaluate \code{math} again. This yields \code{ci_hi}.
\item Add attribute named \code{ci_meta} to the \code{data.table}.
This attribute is a list which contains elements \code{conf_lvl},
\code{conf_method}, and \code{math}.
\item Delete columns \code{std_err}, \code{z}. Rename \code{est} to \code{statistic} and \code{var} to
\code{variance}.
\item Return \code{data.table} with columns
\code{c("statistic", "variance", "ci_lo", "ci_hi")}.
}
}

\examples{

# directadjusting::delta_method_confidence_intervals
dt_1 <- directadjusting::delta_method_confidence_intervals(
  statistics = 0.0,
  variances = 0.1,
  conf_lvl = 0.95,
  conf_method = "identity"
)
# you can also supply your own math for computing the confidence intervals
dt_2 <- directadjusting::delta_method_confidence_intervals(
  statistics = 0.0,
  variances = 0.1,
  conf_lvl = 0.95,
  conf_method = quote(est + z * std_err)
)
stopifnot(
  all.equal(dt_1, dt_2, check.attributes = FALSE)
)
}
