% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_adjusting.R
\name{direct_adjusted_estimates}
\alias{direct_adjusted_estimates}
\title{Direct Adjusted Estimates}
\usage{
direct_adjusted_estimates(stats_dt, stat_col_nms, var_col_nms,
  stratum_col_nms = NULL, adjust_col_nms, conf_lvls, conf_methods,
  weights)
}
\arguments{
\item{stats_dt}{\code{[data.table]} (mandatory, no default)

a \code{data.table} containing estimates and variance estimates of statistics}

\item{stat_col_nms}{\code{[character]} (mandatory, no default)

names of columns in \code{stats_dt} containing estimates (statistics);
\code{NA} statistics values cause also \code{NA} confidence intervals}

\item{var_col_nms}{\code{[character]} (optional, default \code{NULL})
\itemize{
\item if \code{NULL}, no confidence intervals can (will) be computed
\item if \code{character} vector, names of columns in \code{stats_dt} containing variance
estimates of the statistics specified in \code{stat_col_nms} with one-to-one
correspondence; \code{NA} elements in \code{var_col_nms} cause no confidence
intervals to computed for those statistics;
\code{NA} variance estimates in \code{stats_dt} cause \code{NA} confidence intervals;
negative values cause an error; \code{Inf} values cause \code{c(-Inf, Inf)}
intervals with confidence interval method \code{"identity"}, etc.
}}

\item{stratum_col_nms}{\code{[NULL, character]} (optional, default \code{NULL})

names of columns in \code{stats_dt} by which statistics are stratified (and they
should be stratified by these columns after direct adjusting)}

\item{adjust_col_nms}{\code{[character]} (mandatory, no default)

names of columns in \code{stats_dt} by which statistics are currently stratified
and by which the statistics should be adjusted (e.g. \code{"agegroup"})}

\item{conf_lvls}{\code{[numeric]} (mandatory, default \code{0.95})

confidence levels for confidence intervals; you may specify each statistic
(see \code{stat_col_nms}) its own level by supplying a vector of values;
values other than between \code{(0, 1)} cause an error}

\item{conf_methods}{\code{[character]} (mandatory, default \code{"identity"})

method to compute confidence intervals; either one string (to be used for
all statistics) or a vector of strings, one for each element of
\code{stat_col_nms}; use \code{"none"} for statistics for which you do not want
confidence intervals to be calculated (or \code{NA} in \code{var_col_nms})
see \code{\link{confidence_interval_methods}} for supported methods}

\item{weights}{\code{[double, data.table, character]}

the weights need not sum to one as this is ensured internally; you may
supply weights in one of the following ways:
\itemize{
\item \code{double}: a vector of weights can be given when adjusting by exactly
one variable (e.g. age group); the length must match the number of unique
values in the adjusting variable
\item \code{data.table}: with one or more columns with names matching to those
variables that are used to adjust estimates, and one column named
\code{weight}; e.g. \code{data.table(agegroup = 1:3, weight = c(100, 500, 400))}
\item \code{character}: a string specifying one of the weighting schemas integrated
into this package; see \code{\link{integrated_weight_schemas}}; again
this is only allowed when you are adjusting by exactly one variable
}}
}
\description{
Compute direct adjusted estimates from a table of statistics.
}
\examples{

# suppose we have poisson rates that we want to adjust for by age group.
# they are stratified by sex.
library("data.table")
set.seed(1337)

offsets <- rnorm(8, mean = 1000, sd = 100)
baseline <- 100
sex_hrs <- rep(1:2, each = 4)
age_group_hrs <- rep(c(0.75, 0.90, 1.10, 1.25), times = 2)
counts <- rpois(8, baseline * sex_hrs * age_group_hrs)

# raw estimates
my_stats <- data.table(
  sex = rep(1:2, each = 4),
  ag = rep(1:4, times = 2),
  e = counts / offsets
)
my_stats[["v"]] <- my_stats[["e"]] / offsets

# adjusted by age group
my_adj_stats <- direct_adjusted_estimates(
  stats_dt = my_stats,
  stat_col_nms = "e",
  var_col_nms = "v",
  conf_lvls = 0.95,
  conf_methods = "log",
  stratum_col_nms = "sex",
  adjust_col_nms = "ag",
  weights = c(200, 300, 400, 100)
)

}
