% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_directadjusting.R
\docType{package}
\name{directadjusting-package}
\alias{directadjusting}
\alias{directadjusting-package}
\title{directadjusting: Directly Adjusted Estimates}
\description{
Compute estimates and confidence intervals of weighted averages quickly and easily.

\href{https://github.com/FinnishCancerRegistry/directadjusting/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/FinnishCancerRegistry/directadjusting/actions/workflows/R-CMD-check.yaml/badge.svg}}}
}
\section{Recommended installation}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github(
  "FinnishCancerRegistry/directadjusting",
  ref = readline("enter latest tag on github: ")
)
}\if{html}{\out{</div>}}
}

\section{Example}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{library("directadjusting")

# suppose we have poisson rates that we want to adjust for by age group.
# they are stratified by sex.
library("data.table")
set.seed(1337)

offsets <- rnorm(8, mean = 1000, sd = 100)
baseline <- 100
sex_hrs <- rep(1:2, each = 4)
age_group_hrs <- rep(c(0.75, 0.90, 1.10, 1.25), times = 2)
counts <- rpois(8, baseline * sex_hrs * age_group_hrs)

# raw estimates
my_stats <- data.table(
  sex = rep(1:2, each = 4),
  ag = rep(1:4, times = 2),
  e = counts / offsets
)
my_stats[["v"]] <- my_stats[["e"]] / offsets

# adjusted by age group
my_adj_stats <- direct_adjusted_estimates(
  stats_dt = my_stats,
  stat_col_nms = "e",
  var_col_nms = "v",
  conf_lvls = 0.95,
  conf_methods = "log",
  stratum_col_nms = "sex",
  adjust_col_nms = "ag",
  weights = c(200, 300, 400, 100)
)

}\if{html}{\out{</div>}}
}

\section{News}{

\subsection{News for version 0.5.0}{
\subsection{directadjusting::direct_adjusted_estimates}{

\code{directadjusting::delta_method_confidence_intervals} made a lot more
flexible. It now accepts via \code{conf_method} a string, a call, and a list
of calls that produce the desired confidence intervals.
}

\subsection{directadjusting::direct_adjusted_estimates}{

\code{directadjusting::direct_adjusted_estimates} option \code{conf_methods = "boot"}
removed. Only delta method confidence intervals now possible.
}

\subsection{directadjusting::direct_adjusted_estimates}{

\code{directadjusting::direct_adjusted_estimates} now allows for the sake of
convenience to be called with no \code{adjust_col_nms} defined. This results
in no adjusting.
}

}

\subsection{News for version 0.4.0}{
\subsection{directadjusting::direct_adjusted_estimates}{

\code{directadjusting::direct_adjusted_estimates} now correctly uses
the same \code{conf_lvls} and \code{conf_methods} for all statistics when their
length is one.
}

}

\subsection{News for version 0.3.0}{
\subsection{directadjusting}{

Remove deprecated \code{directadjusting::direct_adjusted_estimates}. Use
\code{directadjusting::directly_adjusted_estimates}.
}

}
}

\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.miettinen@cancer.fi} (\href{https://orcid.org/0000-0001-8624-6754}{ORCID})

}
\keyword{internal}
